+++ examples/demo_app/__init__.py
@@ -1,0 +1,1 @@
+"""Bundled demo app example target."""